Refactor: refactoring screen folder structure
