+++ src/navigations/navigation-screen.tsx
@@ -1,8 +1,7 @@
 // Screen
 import { ChangePassword } from 'screens/change-password';
 import { EditProfile } from 'screens/edit-profile';
-import { Home } from 'screens/home';
-import { Inventory } from 'screens/invetory';
+import { Inventory } from 'screens/inventory';
 import { Notifications } from 'screens/notifications';
 import { SignIn } from 'screens/sign-in';
 import { SignUp } from 'screens/sign-up';
