+++ src/navigations/stacks/home-stack.tsx
@@ -1,6 +1,3 @@
-/* eslint-disable react/no-unstable-nested-components */
-/* eslint-disable react-native/no-inline-styles */
-/* eslint-disable prettier/prettier */
 // Modules
 import React, { useEffect } from 'react';
 import { createStackNavigator } from '@react-navigation/stack';
